refactor(api): extract webhooks from client.ts
